fixup! Sema: ignore macro in qualified import
