--- test/ImportResolution/Inputs/StructAndMacro.swift
@@ -1,7 +1,0 @@
-public struct Foo {
-  public var value: Int
-  public init(value: Int) { self.value = value }
-}
-
-@freestanding(expression)
-public macro Foo() -> Int = #externalMacro(module: "NonExistent", type: "NonExistent")
